WFLY-6974 refactor weld transformers to non-deprecated api
